requirements.txt
fastapi
uvicorn
chromadb
pypdf
requests
python-multipart
PyPDF2
pymupdf 
pytesseract 
pillow
python-dotenv
gitpython